backend/requirements.txt
fastapi
uvicorn[standard]
alembic
sqlalchemy
asyncpg
httpx
redis
pydantic-settings
psycopg2-binary
python-jose[cryptography]
rapidfuzz
argon2-cffi
python-multipart
pytest
pytest-asyncio
loguru
slugify
tenacity
openai
requests
openrouteservice